moved AutoSlug into namespaced directory
