+++ scarlet/cms/static/scarlet/source/js/components/autoSlug/AutoSlug.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
-import muiTheme from '../common/muiTheme'
+import { dasherize } from './autoSlugUtils'
 
-import '../../stylesheets/components/AutoSlug.scss'
+import '../../../stylesheets/components/AutoSlug.scss'
 
 class AutoSlug extends Component {
   constructor(props) {
@@ -15,9 +15,17 @@ class AutoSlug extends Component {
     this.setSlugFromSelf = this.setSlugFromSelf.bind(this)
 
     // enable value matching from sourceElem only if initial values match
-    if (this.dasherize(this.props.sourceElem.value) === this.props.slugElem.value) {
+    if (this.shouldEnableMatching()) {
       this.props.sourceElem.onkeyup = this.setSlugFromSource.bind(this)
     }
+  }
+
+  shouldEnableMatching() {
+    return dasherize(this.props.sourceElem.value) === this.props.slugElem.value
+  }
+
+  getSourceValue() {
+    return dasherize(this.props.sourceElem.value)
   }
 
   setSlugFromSource() {
@@ -28,18 +36,13 @@ class AutoSlug extends Component {
 
   setSlugFromSelf(e) {
     // disable value matching from sourceElem
-    this.props.sourceElem.onkeyup = null
+    if (!this.shouldEnableMatching()) {
+      this.props.sourceElem.onkeyup = null
+    }
+
     this.setState({
-      slug: this.dasherize(e.target.value)
+      slug: dasherize(e.target.value)
     })
-  }
-
-  getSourceValue() {
-    return this.dasherize(this.props.sourceElem.value)
-  }
-
-  dasherize(text) {
-    return text.replace(/\s+/g, '-').toLowerCase()
   }
 
   render() {
